Update 04.discussion.md
Added more references to previous discoveries of mutator allele epistasis in yeast + cell lines. Since some of this work involves human cell lines, one could quibble with the claim that ours is the first discovery of mutator epistasis in a vertebrate, so I rewrote to attempt be more precise about what is novel and connect more to other relevant work.